user page styling

+++ front/src/components/ProductDetails.js
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import Cookies from 'js-cookie';
 import axios from "axios";
-import { Link, useHistory, useParams } from "react-router-dom";
+import { useHistory, useParams } from "react-router-dom";
 
 
 const Details = () => {
@@ -10,6 +10,9 @@ const Details = () => {
     const history = useHistory();
     const [product, setProduct] = useState({})
     const [users, setUsers] = useState([])
+    const [open, setOpen] = useState(false)
+    const [img, setImg] = useState(product.picture)
+    const [name, setName] = useState(product.name)
 
     useEffect(() => {
         axios.get(`http://localhost:5000/products/${id}`)
@@ -43,8 +46,19 @@ const Details = () => {
     }
 
     const delProduct = () => {
-        axios.delete(`http://localhost:5000/products/${product._id}`)
-        .then(() => history.push('/'))
+        const c = window.confirm('Are you sure ypu want to delete this product?')
+        if (c) {
+            axios.delete(`http://localhost:5000/products/${product._id}`)
+            .then(() => history.push('/'))
+            .catch(err => console.log(err))
+        } else {}
+    }
+
+    const editProd = () => {
+        axios.patch(`http://localhost:5000/products/${product._id}`, {picture: img, name: name})
+        .then(() => alert('Product successfully updated!'))
+        .then(() => setOpen(false))
+        .catch(err => console.log(err))
     }
 
 
@@ -54,10 +68,15 @@ const Details = () => {
           <div className="product-list">
                <div key={product._id}>
                          <img className="product-pic" src={product.picture} alt="img"/>
+                         {open ? <input defaultValue={product.picture} type='url' onChange={(e) => setImg(e.target.value)}/> : null}
                          <p className="product-name">{product.name}</p>
+                         {open ? <input defaultValue={product.name} type='text' onChange={(e) => setName(e.target.value)}/> : null}
+                         {open ? <button type='submit' onClick={editProd}>OK</button> : null}
+                         {open ? <button onClick={() => setOpen(false)}>CANCEL</button> : null}
                          <p>{'\u2605'} {product.rating}</p>
+                         <p>{product.prize}pln</p>
                          <button onClick={updateCart(product.name)}>ADD TO CART</button>
-                         {users.map(u => u.email === Cookies.get()['user'] && u.role==='admin' ? <Link to={`/${product._id}/edit`}><button>EDIT</button></Link> : null)}
+                         {users.map(u => u.email === Cookies.get()['user'] && u.role==='admin' ? <button onClick={() => setOpen(true)}>EDIT</button> : null)}
                          {users.map(u => u.email === Cookies.get()['user'] && u.role==='admin' ? <button onClick={delProduct}>DELETE</button> : null)}
                </div>
         </div>
